Add files via upload
In this project, I focused on the critical problem of customer churn prediction within the telecom industry. The objective is to understand the reasons behind customers leaving a telecom business. The primary goal is to create a deep learning model (ANN) to accurately predict customer churn. To evaluate the model's effectiveness, I employed performance metrics, including precision, recall, and the F1-score.